Milestone 1 - Implementation of the Homepage
A simple homepage for the user to see once they enter into the site. A background image has been used with an off-white shape placed on top, in which the other elements of the website will be placed over. The name of the system is displayed using a header and two buttons (which will be implemented) are displayed beneath the header, with the intention for the users to click on.