Milestone 2 - Implementation of the booking system
Functionality to the "Create a Booking" button was added, which directs the user to the Booking page. The Booking page has also been created, in which the user is displayed 6 input text fields and one text area, alongside their respective labels. The user is able to enter in information within the text fields and text area. Once the user hits confirm, the information is converted to a JavaScript object, then using JSON, is stored within local storage for the system to use in the future. Additional CSS as also been added to correctly format the booking page.