Milestone 3 - Implementation of the edit function
The ability to search of bookings from the homepage has been added, alongside the ability to edit the bookings already made. To search for a booking, two input text fields are provided which ask for First and Last names, which then searches the local storage for the correct booking and then displays as such. The "Change flight details" button has also been implemented, which takes the user to a different page. The user is displayed the same information as per the "Create a booking" page, but with their data already within the input fields. The user can change any of the these fields and once "Confirm" is pressed, the system will update its database and store the information back into it's local storage.