Milestone 4 - Implementation of in-flight services
Functionality to the "Purchase in-flight services" button has been added, which redirects the user to the "Purchase Inflight Services" page when clicked on. This page displays 5 checkboxes, which the user can click on to purchase the inflight services. When the user clicks confirm, the system will store the data alongside the other booking data and then update the database within the local storage. The changes are then reflected on the "View booking" page, which displays all the booking information.