Milestone 5 - Finalisation of the Booking System
Finalisation of the Booking System. An additional page has been added to show the user that their booking as been completed. When the user clicks on "Confirm" from the "View Booking" page, they are directed to the confirmation page to notify the user that their booking as been confirmed. The User can then return back to the homepage from the "Confirmation" page, by click the "Return Home" button, which then redirects the user back to the homepage. Minor bug fixes were also made, alongside the finalisation of the background image.